requirements.txt
discord.py
spotipy
yt-dlp
mutagen
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib